requirements.txt
torch~=1.11.0
numpy~=1.22.3
transformers~=4.17.0
PyYAML~=5.3.1
dill~=0.4.0